requirements.txt
Django==4.2.6
djangorestframework==3.14.0
drf-spectacular==0.26.5
psycopg2==2.9.9
Pillow==10.0.1
